\name{scRNAseq-package}
\alias{scRNAseq-package}
\alias{scRNAseq}
\alias{fluidigm}
\alias{th2}
\alias{shalek_dc}

\docType{package}
\title{
\packageTitle{scRNAseq}
}
\description{
\packageDescription{scRNAseq}
}
\details{

This package contains a collection of three publicly available single-cell RNA-seq data.
The datasets are used for the evaluation of the scone package.

The dataset \code{fluidigm} contains 65 cells from Pollen et al. (2014), each sequenced at high and low coverage.

The dataset \code{th2} contains 96 T helper cells from Mahata et al. (2014).

The dataset \code{shalek_dc} contains 2397 dendritic cells from Shalek et al. (2014).

See the package vignette for details on the pre-processing of the data.

}
\author{
\packageAuthor{scRNAseq}

Maintainer: \packageMaintainer{scRNAseq}
}
\references{

Pollen, Nowakowski, Shuga, Wang, Leyrat, Lui, Li, Szpankowski, Fowler, Chen, Ramalingam, Sun, Thu, Norris, Lebofsky, Toppani, Kemp II, Wong, Clerkson, Jones, Wu, Knutsson, Alvarado, Wang, Weaver, May, Jones, Unger, Kriegstein, West. Low-coverage single-cell mRNA sequencing reveals cellular heterogeneity and activated signaling pathways in developing cerebral cortex. Nature Biotechnology, 32, 1053-1058 (2014).

Mahata, Zhang, Kolodziejczyk, Proserpio, Haim-Vilmovsky, Taylor, Hebenstreit, Dingler, Moignard, Gottgens, Arlt, McKenzie, Teichmann. Single-Cell RNA Sequencing Reveals T Helper Cells Synthesizing Steroids De Novo to Contribute to Immune Homeostasis. Cell Reports, 7(4): 1130–1142 (2014).

Shalek, Satija, Shuga, Trombetta, Gennert, Lu, Chen, Gertner, Gaublomme, Yosef, Schwartz, Fowler, Weaver, Wang, Wang, Ding, Raychowdhury, Friedman, Hacohen, Park, May, Regev. Single-cell RNA-seq reveals dynamic paracrine control of cellular variation. Nature, 510, 363-369 (2014).

}